django/requirements.txt
asgiref==3.7.2
b2sdk==1.29.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.4.1
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
logfury==1.0.1
packaging==23.2
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.1.0
